feat(nuxt): update nuxt module
