+++ packages/nuxt/src/runtime/app/utils/gqfn.ts
@@ -1,0 +1,1 @@
+export { useSchema as useGQFnSchema } from '@gqfn/core'
