--- packages/nuxt/src/runtime/index.ts
@@ -1,1 +1,0 @@
-export type { RequireOperationPartialData, ResultOf, TypedDocumentNode, VariablesOf } from '@gqfn/core/types'
