+++ packages/nuxt/src/runtime/server/type.ts
@@ -1,0 +1,1 @@
+export type { RequireOperationPartialData, ResultOf, TypedDocumentNode, VariablesOf } from '@gqfn/core'
